Scope 'non-normative' header to individual sections in Introduction (#51)
Currently the "*This section is non-normative*" disclaimer is applied to all of section '1. Introduction', even though the "Introduction" contains several subsections that define terms and contain other non-normative text.

Scope the "non-normative" header to only the sections that are fully non-normative.

In the future we should consider refactoring such that the sub-sections with normative text are pulled out of the Introduction section.