Clarify designMode corner case (#83)
In https://github.com/w3c/edit-context/issues/74#issuecomment-1806318460 we resolved that EditContext should be a no-op when applied to an element that is already editable due to `designMode` being set for the document.

For the most part this was already true per the current spec because EditContext is a no-op when set on a node whose parent element is editable. However there was some ambiguity in the case where EditContext is set on the [documentElement](https://dom.spec.whatwg.org/#document-element). Add a specific check for that case, stating that that element won't be an EditContext editing host if the document has `designMode` set to "on".

Closes https://github.com/w3c/edit-context/issues/74.