Remove codepoint locations struct (#84)
The "codepoint locations" struct is defined but not really referenced throughout the rest of the spec. Its intention was to allow for multiple ranges of cached codepoint rects, but the getters and setters for character bounds spec assume that only one such range exists.

Since it doesn't seem that support for multiple ranges ended up being needed, remove the "codepoint locations" concept so that the setters/getters are consistent with the actual shape of EditContext's data structures.